// types.ts
export interface LifestyleData {
    age: number | string;
    lifestyleFactors: string[];
    familyHistory: string;
}

export interface HealthRisk {
    risk: string;
    explanation: string;
    preventionTips: string[];
    riskLevel: 'High' | 'Medium' | 'Low';
}
export interface PredictiveAnalysis {
    summary: string;
    potentialRisks: HealthRisk[];
}